+++ client/src/ptest/TestsManifest.ts
@@ -8,7 +8,7 @@ import { readFileSync } from 'fs';
 import { Test } from './Test';
 import { Uri, window } from 'vscode';
 import { PddlExtensionContext } from '../../../common/src/PddlExtensionContext';
-import * as fs from 'fs';
+import * as afs from '../../../common/src/asyncfs';
 
 /**
  * Tests manifest
@@ -56,7 +56,7 @@ export class TestsManifest {
 
         var json = JSON.stringify(obj, null, 2);
         try {
-            await fs.promises.writeFile(this.uri.fsPath, json, 'utf8');
+            await afs.writeFile(this.uri.fsPath, json, 'utf8');
         }
         catch(err) {
             window.showErrorMessage(`Error saving test case manifest ${err.name}: ${err.message}`);